requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-modeltranslation==0.18.11
psycopg2-binary==2.9.9
pillow==10.1.0
python-dotenv==1.0.0
dj-database-url==2.1.0
gunicorn==21.2.0
qrcode==7.4.2
social-auth-app-django==5.3.0
django-storages==1.14.2
django-filter==23.3
whitenoise==6.6.0
boto3==1.28.64
django-cleanup==8.0.0
django-compressor==4.4
drf-yasg==1.21.7  
